+++ src/app/components/forms/dynamic-form/terra-dynamic-form.component.ts
@@ -32,14 +32,34 @@ export interface TerraDynamicFormRequestParams
 @Component({
     selector:  'terra-dynamic-form',
     template:  require('./terra-dynamic-form.component.html'),
+    styles:    [require('./terra-dynamic-form.component.scss')],
     providers: [TerraDynamicFormService]
 })
 export class TerraDynamicFormComponent implements OnInit, OnChanges
 {
-    @Input() public inputFormFunctions:TerraDynamicFormFunctionsHandler<any>;
-    @Input() public inputFormFields:Array<TerraFormFieldBase<any>>;
-    @Input() public inputPortletStyle:string;
-    @Input() public inputRequestParams:TerraDynamicFormRequestParams;
+    @Input()
+    public inputFormFunctions:TerraDynamicFormFunctionsHandler<any>;
+
+    @Input()
+    public inputFormFields:Array<TerraFormFieldBase<any>>;
+
+    @Input()
+    public inputPortletStyle:string;
+
+    @Input()
+    public inputRequestParams:TerraDynamicFormRequestParams;
+
+    @Input()
+    public inputHasNoSaveButton:boolean;
+
+    @Input()
+    public inputHasNoResetButton:boolean;
+
+    @Input()
+    public inputIsDisabled:boolean;
+
+    @Input()
+    public inputUsePortlet:boolean = true;
 
     constructor(private _formFieldControlService:TerraFormFieldControlService)
     {
@@ -49,6 +69,10 @@ export class TerraDynamicFormComponent implements OnInit, OnChanges
             htmlMethod: null,
             params:     {}
         };
+
+        this.inputHasNoSaveButton = false;
+        this.inputHasNoResetButton = false;
+        this.inputIsDisabled = false;
     }
 
     public ngOnInit():void
@@ -61,6 +85,8 @@ export class TerraDynamicFormComponent implements OnInit, OnChanges
         {
             this._formFieldControlService.createFormGroup(this.inputFormFields);
             this.inputFormFunctions.formFieldControlService = this._formFieldControlService;
+
+            this.registerValueChange();
         }
     }
 
@@ -69,6 +95,7 @@ export class TerraDynamicFormComponent implements OnInit, OnChanges
         if(changes['inputFormFields'])
         {
             this._formFieldControlService.createFormGroup(this.inputFormFields);
+            this.registerValueChange();
         }
     }
 
@@ -88,5 +115,20 @@ export class TerraDynamicFormComponent implements OnInit, OnChanges
     protected onResetClick():void
     {
         this._formFieldControlService.resetForm();
+    }
+
+    private registerValueChange():void
+    {
+        if(!isNullOrUndefined(this.inputFormFunctions.onValueChangedCallback))
+        {
+            this._formFieldControlService
+                .dynamicFormGroup
+                .valueChanges
+                .debounceTime(1000)
+                .subscribe((value:any) =>
+                {
+                    this.inputFormFunctions.onValueChangedCallback(value);
+                });
+        }
     }
 }