+++ src/app/components/forms/input/number-input/terra-number-input.component.ts
@@ -24,8 +24,13 @@ let nextId:number = 0;
 export class TerraNumberInputComponent extends TerraInputComponent
 {
     /**
+     * @description a unique string identifier for the specific input instance.
+     */
+    private _id:string;
+
+    /**
      * @deprecated inputValue is deprecated. It will be removed in one of the upcoming releases. Please use ngModel instead.
-     * */
+     */
     @Input()
     public set inputValue(v:number)
     {
@@ -33,11 +38,6 @@ export class TerraNumberInputComponent extends TerraInputComponent
 
         this.value = v;
     }
-
-    /**
-     * @description a unique string identifier for the specific input instance.
-     */
-    private _id:string;
 
     constructor()
     {