+++ src/app/components/forms/suggestion-box/data/terra-suggestion-box.interface.ts
@@ -1,7 +1,7 @@
 export interface TerraSuggestionBoxValueInterface
 {
     value:any;
-    //i18n from locale_de and locale_en
+    // i18n from locale_de and locale_en
     caption:any;
     icon?:string;
     active?:boolean;