+++ src/app/components/layouts/taglist/terra-taglist.component.ts
@@ -1,7 +1,6 @@
 import {
     Component,
-    Input,
-    OnInit
+    Input
 } from '@angular/core';
 import { TerraTagInterface } from '../tag/data/terra-tag.interface';
 
@@ -10,15 +9,8 @@ import { TerraTagInterface } from '../tag/data/terra-tag.interface';
     styles:   [require('./terra-taglist.component.scss')],
     template: require('./terra-taglist.component.html')
 })
-export class TerraTaglistComponent implements OnInit
+export class TerraTaglistComponent
 {
-    @Input() inputTagList:Array<TerraTagInterface>;
-
-    constructor()
-    {
-    }
-
-    ngOnInit()
-    {
-    }
+    @Input()
+    public inputTagList:Array<TerraTagInterface>;
 }