+++ src/app/components/pager/data/terra-pager.data.ts
@@ -1,0 +1,11 @@
+export class TerraPagerData
+{
+    public page:number;
+    public totalsCount:number;
+    public isLastPage:boolean;
+    public lastPageNumber:number;
+    public firstOnPage:number;
+    public lastOnPage:number;
+    public itemsPerPage:number;
+    public pagingUnit:string; // Label for pager items
+}