+++ src/app/components/split-view/multi/data/terra-multi-split-view.config.ts
@@ -24,7 +24,7 @@ export class TerraMultiSplitViewConfig
         if(view.parameter)
         {
             console.warn(
-                'Property \'parameter\' is deprecated. It will be removed in one of the upcoming releases. Please use \'inputs\' instead.')
+                'Property \'parameter\' is deprecated. It will be removed in one of the upcoming releases. Please use \'inputs\' instead.');
         }
 
         // TODO: setTimeout can be removed, if it is guaranteed that change detection is fired when adding a new view
@@ -68,7 +68,7 @@ export class TerraMultiSplitViewConfig
                                 (child.inputs && view.inputs && CircularJSON.stringify(child.inputs) === CircularJSON.stringify(view.inputs)) ||
                                 (child.name === view.name);
 
-                            if(hasSameParameter && child.module.ngModule == view.module.ngModule)
+                            if(hasSameParameter && child.module.ngModule === view.module.ngModule)
                             {
                                 view = child;
                                 viewExist = true;
@@ -97,7 +97,7 @@ export class TerraMultiSplitViewConfig
 
         let parent:TerraMultiSplitViewInterface = view.parent;
 
-        let viewIndex:number = parent.children.findIndex((elem) => elem === view);
+        let viewIndex:number = parent.children.findIndex((elem:TerraMultiSplitViewInterface) => elem === view);
 
         if(viewIndex >= 0)
         {