+++ src/app/components/tree/checkbox-tree/terra-checkbox-tree.component.ts
@@ -8,6 +8,7 @@ import {
 } from '@angular/core';
 import { TerraBaseTreeComponent } from '../base/terra-base-tree.component';
 import { TerraCheckboxLeafInterface } from '../leaf/terra-checkbox-leaf.interface';
+import { TerraCheckboxTreeLeafState } from './data/terra-checkbox-tree-leaf-state';
 
 @Component({
     selector: 'terra-checkbox-tree',
@@ -19,37 +20,34 @@ export class TerraCheckboxTreeComponent extends TerraBaseTreeComponent implement
     /**
      * @description current level leaf list
      */
-    @Input() inputLeafList:Array<TerraCheckboxLeafInterface>;
+    @Input()
+    public inputLeafList:Array<TerraCheckboxLeafInterface>;
 
     /**
      * @description leafs one level higher than current leaf
      */
-    @Input() inputParentLeafList:Array<TerraCheckboxLeafInterface>;
+    @Input()
+    public inputParentLeafList:Array<TerraCheckboxLeafInterface>;
 
     /**
      * @description complete leaf list for better and faster searching
      */
-    @Input() inputCompleteLeafList:Array<TerraCheckboxLeafInterface>;
+    @Input()
+    public inputCompleteLeafList:Array<TerraCheckboxLeafInterface>;
 
     /**
      * @description get the current selected leaf list
      */
-
-    @Output() valueChange:EventEmitter<TerraCheckboxLeafInterface> = new EventEmitter<TerraCheckboxLeafInterface>();
+    @Output()
+    public valueChange:EventEmitter<TerraCheckboxLeafInterface> = new EventEmitter<TerraCheckboxLeafInterface>();
 
     public selectedLeafList:Array<TerraCheckboxLeafInterface> = [];
-
-    constructor()
-    {
-        super();
-    }
 
     /**
      * @description event which is triggered when any checkbox is clicked
      * @param event
      * @param leaf
      */
-
     private onCheckboxValueChange(event:boolean, leaf:TerraCheckboxLeafInterface):void
     {
         leaf.checkboxChecked = event;
@@ -59,7 +57,7 @@ export class TerraCheckboxTreeComponent extends TerraBaseTreeComponent implement
         this.valueChange.emit(leaf);
     }
 
-    private resetIndeterminateLeafState(leaf:TerraCheckboxLeafInterface)
+    private resetIndeterminateLeafState(leaf:TerraCheckboxLeafInterface):void
     {
         // reset the isIndeterminate flag on every state change
         leaf.isIndeterminate = false;
@@ -69,7 +67,6 @@ export class TerraCheckboxTreeComponent extends TerraBaseTreeComponent implement
      * @description
      * @param leaf
      */
-
     private recursiveAddLeafToList(leaf:TerraCheckboxLeafInterface):void
     {
         if(leaf.checkboxChecked)
@@ -78,7 +75,7 @@ export class TerraCheckboxTreeComponent extends TerraBaseTreeComponent implement
         }
         else
         {
-            let leafIndex = this.selectedLeafList.indexOf(leaf);
+            let leafIndex:number = this.selectedLeafList.indexOf(leaf);
 
             this.selectedLeafList.splice(leafIndex, 1);
         }
@@ -117,7 +114,7 @@ export class TerraCheckboxTreeComponent extends TerraBaseTreeComponent implement
         if(leaf.leafParent)
         {
             let parentLeaf:TerraCheckboxLeafInterface = leaf.leafParent;
-            let parentLeafState:ParentLeafState = this.getParentLeafState(parentLeaf);
+            let parentLeafState:TerraCheckboxTreeLeafState = this.getParentLeafState(parentLeaf);
 
             // All checkboxes on this leaf level are checked
             if(parentLeafState.allChildrenAreChecked)
@@ -151,9 +148,9 @@ export class TerraCheckboxTreeComponent extends TerraBaseTreeComponent implement
         leaf.checkboxChecked = checkboxChecked;
     }
 
-    private getParentLeafState(leaf:TerraCheckboxLeafInterface):ParentLeafState
+    private getParentLeafState(leaf:TerraCheckboxLeafInterface):TerraCheckboxTreeLeafState
     {
-        let parentLeafState:ParentLeafState = new ParentLeafState();
+        let parentLeafState:TerraCheckboxTreeLeafState = new TerraCheckboxTreeLeafState();
 
         for(let subLeaf of leaf.subLeafList)
         {
@@ -192,13 +189,13 @@ export class TerraCheckboxTreeComponent extends TerraBaseTreeComponent implement
         }
     }
 
-    ngOnInit()
+    public ngOnInit():void
     {
         super.ngOnInit();
         this.appendParentsToLeafList(this.inputLeafList);
     }
 
-    ngOnChanges(changes:SimpleChanges):void
+    public ngOnChanges(changes:SimpleChanges):void
     {
         if (changes['inputLeafList'])
         {
@@ -218,9 +215,4 @@ export class TerraCheckboxTreeComponent extends TerraBaseTreeComponent implement
     }
 }
 
-export class ParentLeafState
-{
-    allChildrenAreChecked:boolean = true;
-    noChildrenAreChecked:boolean = true;
-    isIndeterminate:boolean = false;
-}
+