+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 ### Bug Fixes
 * **terra-multi-split-view**  fix issue when view inputs contain circular references
 
+<a name="2.1.61-zlk"></a>
+# 2.1.61-zlk (23.03.2018)
+
+### Feature
+* **terra-tag** add Styling for menu entry tag in terra-tag component
+* **terra-dynamic-form** new component to pick a color 
+* **terra-dynamic-form** callback function for value changed
+* **terra-dynamic-form** added input for portlet usage in <code>TerraDynamicSwitch</code>, default is <code>true</code>
+* **terra-dynamic-switch** 
+	- added input for portlet usage, default is <code>true</code>
+	- using portlet when type is <code>TerraFormFieldVerticalContainer</code>
+
 <a name="2.1.60-zlk"></a>
 # 2.1.60-zlk (22.03.2018)